backend/requirements.txt
# Core Dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5

# Database
sqlalchemy>=1.4.0
asyncpg>=0.24.0
alembic>=1.7.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
cryptography>=3.4.7
bcrypt>=3.2.0

# Monitoring and Logging
psutil>=5.8.0
prometheus-client>=0.11.0
prometheus_fastapi_instrumentator>=5.7.1
structlog>=21.1.0

# Backup and Storage
python-dotenv>=0.19.0
aiofiles>=0.7.0
python-magic>=0.4.24

# Email
aiosmtplib>=1.1.4
email-validator>=1.1.3

# Cache
redis>=4.0.0
aioredis>=2.0.0

# External Services
aiohttp>=3.8.0
httpx>=0.23.0
websockets>=10.0

# Testing
pytest>=6.2.5
pytest-asyncio>=0.15.1
pytest-cov>=2.12.1
pytest-mock>=3.6.1
pytest-env>=0.6.2

# Development
black>=21.7b0
isort>=5.9.3
flake8>=3.9.2
mypy>=0.910
pre-commit>=2.15.0

# Code quality and testing
mypy>=0.910
flake8>=3.9.0
pytest>=6.2.5
coverage>=6.2
black>=21.12b0

# Collaboration
websockets>=10.0
asyncio>=3.4.3

# Development tools
pytest-asyncio>=0.16.0
pytest-cov>=3.0.0
pytest-mock>=3.6.1

# Documentation
mkdocs>=1.2.0
mkdocs-material>=8.0.0
mkdocstrings>=0.15.0

# Utilities
python-dateutil>=2.8.2
pytz>=2021.1
tenacity>=8.0.1
tqdm>=4.62.0

# AI/ML dependencies
openai>=1.66.5
anthropic>=0.5.0
google-generativeai>=0.3.0
groq>=0.4.0
openai-agents>=0.0.6

# Testing and development
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
black>=23.10.1
isort>=5.12.0
mypy>=1.6.1
flake8>=6.1.0

# Utilities
requests>=2.31.0
python-multipart>=0.0.6
aiohttp>=3.9.1
tqdm>=4.66.1

# Monitoring and logging
prometheus-client>=0.19.0
structlog>=23.2.0

# New dependencies for improvements
opentelemetry-api>=1.11.1
opentelemetry-sdk>=1.11.1
opentelemetry-instrumentation-fastapi>=0.30b1
opentelemetry-exporter-jaeger>=1.11.1
opentelemetry-instrumentation-redis>=0.30b1
opentelemetry-instrumentation-sqlalchemy>=0.30b1
strawberry-graphql>=0.96.0
jwt>=1.3.1